consolidate and simplify covenant JS examples
